requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
markdown>=3.5.0
gunicorn>=21.2.0
networkx>=3.0
groq>=0.9.0
